+++ src/tech/java/net/infstudio/inspiringworld/tech/common/world/SourceBombExplosion.java
@@ -47,37 +47,37 @@ public class SourceBombExplosion extends Explosion {
     @Override
     public void doExplosionA() {
         float f3 = this.explosionSize * 2.0F;
-        int k1 = MathHelper.floor_double(this.explosionX - (double) f3 - 1.0D);
-        int l1 = MathHelper.floor_double(this.explosionX + (double) f3 + 1.0D);
-        int i2 = MathHelper.floor_double(this.explosionY - (double) f3 - 1.0D);
-        int i1 = MathHelper.floor_double(this.explosionY + (double) f3 + 1.0D);
-        int j2 = MathHelper.floor_double(this.explosionZ - (double) f3 - 1.0D);
-        int j1 = MathHelper.floor_double(this.explosionZ + (double) f3 + 1.0D);
+        int k1 = MathHelper.floor_double(this.explosionX - f3 - 1.0D);
+        int l1 = MathHelper.floor_double(this.explosionX + f3 + 1.0D);
+        int i2 = MathHelper.floor_double(this.explosionY - f3 - 1.0D);
+        int i1 = MathHelper.floor_double(this.explosionY + f3 + 1.0D);
+        int j2 = MathHelper.floor_double(this.explosionZ - f3 - 1.0D);
+        int j1 = MathHelper.floor_double(this.explosionZ + f3 + 1.0D);
         List<Entity> list = this.worldObj.getEntitiesWithinAABBExcludingEntity(this.exploder,
-            new AxisAlignedBB((double) k1, (double) i2, (double) j2, (double) l1, (double) i1, (double) j1));
+            new AxisAlignedBB(k1, i2, j2, l1, i1, j1));
         net.minecraftforge.event.ForgeEventFactory.onExplosionDetonate(this.worldObj, this, list, f3);
         Vec3d vec3d = new Vec3d(this.explosionX, this.explosionY, this.explosionZ);
 
         for (Entity entity : list) {
 
             if (!entity.isImmuneToExplosions()) {
-                double d12 = entity.getDistance(this.explosionX, this.explosionY, this.explosionZ) / (double) f3;
+                double d12 = entity.getDistance(this.explosionX, this.explosionY, this.explosionZ) / f3;
 
                 if (d12 <= 1.0D) {
                     double d5 = entity.posX - this.explosionX;
-                    double d7 = entity.posY + (double) entity.getEyeHeight() - this.explosionY;
+                    double d7 = entity.posY + entity.getEyeHeight() - this.explosionY;
                     double d9 = entity.posZ - this.explosionZ;
-                    double d13 = (double) MathHelper.sqrt_double(d5 * d5 + d7 * d7 + d9 * d9);
+                    double d13 = MathHelper.sqrt_double(d5 * d5 + d7 * d7 + d9 * d9);
 
                     if (d13 != 0.0D) {
                         d5 = d5 / d13;
                         d7 = d7 / d13;
                         d9 = d9 / d13;
-                        double d14 = (double) this.worldObj.getBlockDensity(vec3d, entity.getEntityBoundingBox());
+                        double d14 = this.worldObj.getBlockDensity(vec3d, entity.getEntityBoundingBox());
                         double d10 = (1.0D - d12) * d14;
                         entity.attackEntityFrom(
                             new EntityDamageSource("explosion", exploder).setDifficultyScaled().setExplosion(),
-                            (float) ((int) ((d10 * d10 + d10) / 2.0D * 7.0D * (double) f3 + 1.0D)));
+                            ((int) ((d10 * d10 + d10) / 2.0D * 7.0D * f3 + 1.0D)));
                         double d11 = 1.0D;
 
                         if (entity instanceof EntityLivingBase) {
